Initial work on Distinct Command Classes
Each Insteon device has its own schema of messages

Some of this schema is standardized across all devices,
some is slightly different, and some is wholly different.

The schema is most consistent within a device category.

Adding support for all possible message schemas and
commands in a single file would be confusing and hard to
maintain.

One option is to split out the differences based on the
device categories, this is kind of what misterhouse did.
This gets complicated though becuase either the classes
within a device have to change dynamically, or the user
is expected to select and use the correct class.

Instead, since the initial goal of this project is to
provide only the most basic support for common commands
the best way to go is likely to use a single file.

This branch is saved as a demonstration of a means for
moving things into distinct classes.