Add User_Links; Finish Export_Link Function; Add PLM as a device of itself
User_Links are meant to be the user defined links.  Using the export_links
function, a user can import the links which already exist on a device.

Fixed a bug where the PLM would not return itself if asked to get a device
by ID and the ID was the PLM.