Bad Fix for a PLM Oddity
I had a device that was repeatedly causing the PLM to emit 025CXXXXXXX messages.
There is no schema for such messages and it appeared that they were nothing more than 0250XXXXXX messages with the single nibble changed.
I have seen similar things in the past but not this repeatable, nearly every message from the device contained the error.
This was repeatable over many days.
The device was sending messages at the full hop count.
Unplugging and moving the device closer to the modem seems to have solved it.

This particular patch was a bodge and solved some but not all of the issues. Saving for posterity and potentially if it may ever be needed again.